Upgrade iced (v0.13.0 -> v0.14.0) and other dependencies
This is a large change.

`iced_aw` is not yet compatible. Ref:
- https://github.com/iced-rs/iced_aw/issues/365
- https://github.com/iced-rs/iced_aw/issues/364

Instead of waiting, it's been removed. We only used it for a
`number_input` widget. Life is not too terrible even without it.

As part of this iced v0.13.0 -> v0.14.0, we've also adjusted the
styling for the secondary button. The new defaults for this button for
the Light theme are lead to ugly low-contrast gray buttons.